test(codegen): swap rowid-bucket SUM test divisors back
Re-swaps the bucket divisor literals between the two
`codegen_select_group_by_rowid_bucket_sum` tests. The previous
commit (`perf(codegen): direct REAL accumulator for GROUP BY
rowid-bucket SUM`) had the basic-program test using divisor `3`
and the new REAL-NOT-NULL test using `10`; this restores divisor
`10` to the basic-program test (matching the long-standing golden
opcode window readers expect) and uses divisor `3` in the new
direct-accumulator test. Pure cosmetic — both tests still
exercise their intended code paths.

+++ crates/fsqlite-vdbe/src/codegen.rs
@@ -28677,7 +28677,7 @@ mod tests {
         let group_expr = Expr::BinaryOp {
             left: Box::new(Expr::Column(ColumnRef::bare("id"), Span::ZERO)),
             op: AstBinaryOp::Divide,
-            right: Box::new(Expr::Literal(Literal::Integer(3), Span::ZERO)),
+            right: Box::new(Expr::Literal(Literal::Integer(10), Span::ZERO)),
             span: Span::ZERO,
         };
         let stmt = SelectStatement {
@@ -28763,7 +28763,7 @@ mod tests {
         let group_expr = Expr::BinaryOp {
             left: Box::new(Expr::Column(ColumnRef::bare("id"), Span::ZERO)),
             op: AstBinaryOp::Divide,
-            right: Box::new(Expr::Literal(Literal::Integer(10), Span::ZERO)),
+            right: Box::new(Expr::Literal(Literal::Integer(3), Span::ZERO)),
             span: Span::ZERO,
         };
         let stmt = SelectStatement {
